# [opt] Optimization for short circuit of CompoundPred